% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ClearanceParams}
\alias{ClearanceParams}
\title{Clearance Parameters}
\usage{
ClearanceParams(rate = Param(0.01), abx = Param(1, 0), ever_abx = Param(1, 0))
}
\arguments{
\item{rate}{base rate of clearance}

\item{abx}{effect of antibiotics on clearance}

\item{ever_abx}{effect of ever having taken antibiotics on clearance}
}
\value{
A list of parameters for clearance.
}
\description{
Clearance Parameters
}
\examples{
ClearanceParams()
}
