% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CodeToEvent}
\alias{CodeToEvent}
\title{Convert coded events to string events}
\usage{
CodeToEvent(x)
}
\arguments{
\item{x}{A vector of integers}
}
\value{
A vector of strings
}
\description{
Convert coded events to string events
}
\examples{
CodeToEvent(c(-1:19, 21:23, 31:33, -999))
}
