% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EventToCode}
\alias{EventToCode}
\title{Convert string events to coded events}
\usage{
EventToCode(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
A vector of integers
}
\description{
Convert string events to coded events
}
\examples{
EventToCode(c("admission", "discharge", "postest", "negtest"))
EventToCode(c("abxon", "abxoff", "isolon", "isoloff"))
}
