% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{LinearAbxAcquisitionParams}
\alias{LinearAbxAcquisitionParams}
\title{Linear Antibiotic Acquisition Parameters}
\usage{
LinearAbxAcquisitionParams(
  base = Param(0.001),
  time = Param(1, 0),
  mass = Param(1, 1),
  freq = Param(1, 1),
  col_abx = Param(1, 0),
  suss_abx = Param(1, 0),
  suss_ever = Param(1, 0)
)
}
\arguments{
\item{base}{The base rate of acquisition (\eqn{\beta_0}).}

\item{time}{The time effect on acquisition (\eqn{\beta_\mathrm{time}}).}

\item{mass}{The mass action effect on acquisition (\eqn{\beta_\mathrm{mass}}).}

\item{freq}{The frequency effect on acquisition (\eqn{\beta_\mathrm{freq}}).}

\item{col_abx}{The effect for colonized individuals on antibiotics (\eqn{\beta_{\bullet℞}}).}

\item{suss_abx}{The effect on susceptible individuals currently on antibiotics (\eqn{\beta_{\circ℞}}).}

\item{suss_ever}{The effect on susceptible individuals ever being on antibiotics (\eqn{\beta_{\circ^*℞}}).}
}
\value{
A list of parameters for acquisition.
}
\description{
Acquisition parameters for LinearAbxModel and LinearAbxModel2.
}
\details{
Notation: States are represented by circles - susceptible (\eqn{\circ}),
colonized (\eqn{\bullet}), latent (\eqn{\circ^*}). The symbol ℞ represents
antibiotic effects.

The model for this acquisition model is given by

\deqn{
P(\mathrm{Acq(t)}) =
    \left[e^{\beta_\mathrm{time}(t-t_0)}\right]\\
\left\{e^{\beta_0}
    \left[
        \left(
          \frac{\beta_\mathrm{freq}}{P(t)}+(1 - e^{\beta_\mathrm{freq}})
        \right)
        e^{\beta_\mathrm{mass}}\left(
            (N_{\bullet}(t) - N_{\bullet℞}(t)) + e^{\beta_{\bullet℞}}N_{\bullet℞}(t)
            \right)
        + 1 - e^{\beta_\mathrm{mass}}
        \right]
    \right\}\\
 \left[
     N_{\circ}(t) - N_{\circ^*℞}(t) +
     e^{\beta_{\circ^*℞}}
     \left(
       \left(
         N_{\circ^*℞}(t) - N_{\circ℞}(t)
       \right)
       + N_{\circ℞}(t)e^{\beta_{\circ℞}}
     \right)
 \right]
}{
 P(Acq(t)) = exp(beta_time*(t-t0)) * {
   exp(beta_0) * [
     (beta_freq/P(t)
   + (1-exp(beta_freq))) * exp(beta_mass) * ((N_c(t) - N_ca(t)) + exp(beta_col_abx)*N_ca(t)) + 1 - exp(beta_mass)
   ]
 }
 * [
   N_S(t) - N_E(t) + exp(beta_suss_ever)*((E_i(t)-A_i(t)) + A_i(t)*exp(beta_suss_abx))
 ]
}
where P(Acq(t)) is the acquisition probability at time t, with effects from time (\eqn{\beta_\mathrm{time}}),
mass action (\eqn{\beta_\mathrm{mass}}), frequency dependence (\eqn{\beta_\mathrm{freq}}),
colonized individuals on antibiotics (\eqn{\beta_{\bullet℞}}),
and susceptible individuals currently (\eqn{\beta_{\circ℞}}) or ever (\eqn{\beta_{\circ^*℞}}) on antibiotics.
}
\examples{
LinearAbxAcquisitionParams()
}
