% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{OutOfUnitInfectionParams}
\alias{OutOfUnitInfectionParams}
\title{Out of Unit Infection Parameters}
\usage{
OutOfUnitInfectionParams(
  acquisition = Param(0.05),
  clearance = Param(0.01),
  progression = Param(0)
)
}
\arguments{
\item{acquisition}{Rate of acquisition of the disease moving into latent state.}

\item{clearance}{Rate of clearance of the disease moving into uncolonized state.}

\item{progression}{Rate of progression of the disease moving into colonized state.}
}
\value{
A list of parameters for out of unit infection.
}
\description{
Out of Unit Infection Parameters
}
\examples{
OutOfUnitInfectionParams()
}
