% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ParamWRate}
\alias{ParamWRate}
\title{Specify a random testing parameter with a rate.}
\usage{
ParamWRate(param = Param(), rate = Param())
}
\arguments{
\item{param}{Values for the positive rate of the test.}

\item{rate}{Values for the rate of the test.}
}
\value{
A list of with param and rate.
}
\description{
Specify a random testing parameter with a rate.
}
\examples{
ParamWRate(Param(0.5, 0), rate = Param(1, 0))
}
