% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{RandomTestParams}
\alias{RandomTestParams}
\alias{ClinicalTestParams}
\title{Random Testing Parameter Set}
\usage{
RandomTestParams(
  uncolonized = ParamWRate(Param(0.5, 0), Param(1, 0)),
  colonized = ParamWRate(Param(0.5, 0), Param(1, 0)),
  latent = ParamWRate(Param(0), Param(0))
)

ClinicalTestParams(
  uncolonized = ParamWRate(Param(0.5, 0), Param(1, 0)),
  colonized = ParamWRate(Param(0.5, 0), Param(1, 0)),
  latent = ParamWRate(Param(0), Param(0))
)
}
\arguments{
\item{uncolonized}{Testing when the individual is uncolonized.}

\item{colonized}{Testing when the individual is colonized.}

\item{latent}{Testing when the individual is latent.}
}
\value{
list of parameters for random testing.
}
\description{
Random Testing Parameter Set
}
\section{Functions}{
\itemize{
\item \code{ClinicalTestParams()}: Clinical Test Parameters Alias

}}
\examples{
RandomTestParams()
}
