% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.data.R
\docType{data}
\name{simulated.data}
\alias{simulated.data}
\title{Simulated Transmission Data}
\format{
An object of class \code{data.frame} with 8360 rows and 5 columns.
}
\usage{
simulated.data
}
\description{
This data set contains simulated transmission data for a hypothetical infectious disease. The data set contains the following columns:
}
\details{
\itemize{
\item \code{facility}: The facility where the event occurred.
\item \code{unit}: The unit within the facility where the event occurred.
\item \code{time}: The time at which the event occurred.
\item \code{patient}: The patient involved in the event.
\item \code{type}: The type of event.
}
}
\keyword{datasets}
