% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.data.R
\docType{data}
\name{simulated.data_sorted}
\alias{simulated.data_sorted}
\title{Simulated Transmission Data (Sorted)}
\format{
An object of class \code{data.frame} with 8360 rows and 5 columns.
}
\usage{
simulated.data_sorted
}
\description{
This is a sorted version of the simulated transmission data. The data is sorted by time, then by patient.
See \code{\link{simulated.data}} for column descriptions.
}
\keyword{datasets}
