% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{SurveillanceTestParams}
\alias{SurveillanceTestParams}
\title{Surveillance Test Parameters}
\usage{
SurveillanceTestParams(
  colonized = Param(init = 0.8, weight = 1),
  uncolonized = Param(init = 1e-10, weight = 0),
  latent = Param(init = 0, weight = 0)
)
}
\arguments{
\item{colonized}{Also known as the true positive rate for a two state model.}

\item{uncolonized}{Also known as the false positive rate for a two state model.}

\item{latent}{The rate of positive tests when the individual is in the (optional) latent state.}
}
\value{
A list of parameters for surveillance testing.
}
\description{
Specify the rates of positive tests for each state of the model.
}
\examples{
SurveillanceTestParams()
}
