% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ClearanceParams}
\alias{ClearanceParams}
\title{Clearance Parameters}
\usage{
ClearanceParams(rate = Param(0.01), abx = Param(1, 0), ever_abx = Param(1, 0))
}
\arguments{
\item{rate}{Base rate of clearance (\eqn{\gamma_0})}

\item{abx}{Effect of antibiotics on clearance (\eqn{\gamma_{℞}})}

\item{ever_abx}{Effect of ever having taken antibiotics on clearance (\eqn{\gamma_{\circ^*℞}})}
}
\value{
A list of parameters for clearance.
}
\description{
Clearance Parameters
}
\examples{
ClearanceParams()
}
