% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ProgressionParams}
\alias{ProgressionParams}
\title{Progression Parameters}
\usage{
ProgressionParams(
  rate = Param(0.01),
  abx = Param(1, 0),
  ever_abx = Param(1, 0)
)
}
\arguments{
\item{rate}{Base progression rate (\eqn{\delta_0})}

\item{abx}{Effect of current antibiotics on progression (\eqn{\delta_{℞}})}

\item{ever_abx}{Effect of ever having taken antibiotics on progression (\eqn{\delta_{\circ^*℞}})}
}
\value{
A list of parameters for progression.
}
\description{
Progression Parameters
}
\examples{
ProgressionParams()
}
